back-end/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
gunicorn==20.1.0
psycopg2==2.9.3
docker==5.0.3
python-dotenv==0.19.2
